main reqs met, not responsive or good looking tho
