+++ src/assignments/project-1/components/card.js
@@ -1,4 +1,3 @@
-
 import React, { Component } from 'react'
